+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/IncorrectSignaturesConfigKitTest.groovy
@@ -40,4 +40,35 @@ class IncorrectSignaturesConfigKitTest extends AbstractKitTest {
 
     }
 
+    def "Check violation detection without cache task 2"() {
+        setup:
+        build """
+            plugins {
+                id 'java'
+                id 'ru.vyarus.animalsniffer'
+            }
+
+            animalsniffer {
+                failWithoutSignatures = false
+                ignoreFailures = true
+            }
+
+            repositories { mavenCentral()}
+            dependencies {
+                signature 'org.codehaus.mojo.signature:java16-sun:1.0'
+                implementation 'org.slf4j:slf4j-api:1.7.25'
+            }
+
+        """
+        fileFromClasspath('src/main/java/invalid/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java')
+//        debug()
+
+        when: "run task"
+        BuildResult result = run('check')
+
+        then: "task successful"
+        result.task(':check').outcome == TaskOutcome.UP_TO_DATE
+        result.task(':animalsnifferMain').outcome == TaskOutcome.SKIPPED
+    }
+
 }